requirements.txt
jsonschema
simplejson
requests
ConfigParser
behave
allure-behave
PyHamcrest